Add the Bioperl Species.pm module for patching
This module fails on the test_species_patch.ncr file because it contains
an OS lines spanning multiple lines and the first line contains an open
parenthesis. Bioperl creates a regex out of that line, and since the
parenthesis is unclosed, the regex is invalid. The module will be patched
locally in the next commit.