Implement some logic for extracting data
This covers the initial basic logic of extracting data from EMBL files
from ensembl. Currently, this has lots of other changes, such as require
attrs, and a different interface to creating new RNAcentralEntry's.
These changes are probably not going to stay but this is how I am
thinking about organizing the code, generally speaking.

In addition, I have several tests that show what the parser should
extract and store. Hopefully these are useful across the range of
organisms we have to deal with.