Improve ensembl parsing
This does a better job at parsing out the data from the EMBL files that
ensembl produces. This seems to do the right thing based upon what I
understand of the database. However, it is really important th I check
over what I think the database is supposed to store.

The next task is to properly integrate this code with the rest of the
import code base so we can easily export data.