Work on refactor Ensembl import
The code needs to be refactored to allow the duplication of importing
data when dealing with GENCODE data. The key change is that we should
now not assume that each feature will produce a single entry to write.
It may now produce more than one entry. This allows the GENCODE writing
to work. This code is not yet tested or working but shows the general
outline of how things will work.

I've also moved the code around to reflect the hierarchy of the code as
well.