JSON encode all notes
This will parse the notes and then JSON encode them into as an object.
The values in the JSON encoded object will be a list of strings. This is
because in theory there may be multiple xrefs to a single DB so we
always use a list to provide a consistent representation of the data.