Add transcript id to description
This will add the transcript id as part of the description. Right now
this is minimally tested in terms of how good it is. But we use it. In
addition, I've added a couple asserts to ensure that we don't write
empty data in parts of the description. I'm not sure if this will crash
things but it is useful to have.