Change to correctly detecting if entry is GENCODE
I didn't change the logic to deal with the fact that the db_xrefs entry
is a JSON encoded string. We now decode the string to get the required
data. In addition, I fixed a couple other bugs and added the ability to
have a gencode specific description. The description is pretty general
now but it is there at least.