Change to using an atomic_file
A new part of luigi is the atomic_file file. This target is just a file
that will only exist once writing completes correctly. To use it I have
to change several things. Notably, the exists method has to be changed,
as well as the writing and where the files are opened and closed. These
changes are driven by the fact that the atomic_file is not a real target
(does not have it's own exists method) and we have to do more work to
wrap it and ensure the files are closed correctly.