Improve rna_type handling
Ensembl and GENCODE use several rna_types that are not part of the
standard nomenclature. To deal with this I've added some logic to
normalize some things and skip others. The GENCODE naming can be read
about at:

http://www.gencodegenes.org/gencode_biotypes.html

My changes are to:

- TEC, retained_intron, processed_transcript
    Ignoring them as if they are not functional (probably) ncRNA
- macro_lncRNA, sense_intronic, sense_overlapping, bidirectional_promoter_lncRNA
    lncRNA
- non_coding, known_ncrna, 3prime_overlapping_ncRNA
    ncRNA
- Mt_rRNA, Mt_tRNA
    rRNA, tRNA respectively

They have scaRNA which is not part of the standard naming. However it
looks like a perfectly valid name and has Rfam families and a nice wiki
page. We are keeping it for now.