Improve pseudogene detection
A transcript is treated as a pseudogene if the gene it comes from has
the word 'pseudogene' in it's note qualifier, slightly hacky but seems
to work nicely. It requires that the word exist and should not match
things like nonpseudogene (though it will probably match
non-pseudogene). This improves the detection of things called
pseudogenes.

I've
added a bunch of code for storing the gene info in an object that may
prove unneeded but for now it is worthwhile to have such a container. In
addition most of the generic_tests are failing and need to be fixed
shortly.