Normalize taxonomy
Previously this would use the lineage in the EMBL file. This can and
does in the case of human differ from the one used elsewhere. It has the
potential of causing issues in our lineage displays as the same organism
will be placed into different trees confusing the user.

This will now use the taxonomic information provided by EBI's taxonomy
service (https://www.ebi.ac.uk/ena/browse/taxonomy-service). In the case
of human this sets the taxonomy to the one used in our site. This is
useful to ensure we have a consistent one throughout the site. In
addition, it would probably be a good idea to enforce this more
generally.

As part of the change I've moved where some of the standard annotations
are computed so they are always available.