Refactor and deal with lincRNA correctly
This is a bit of a refactoring of how data is stored and the interfaces
to methods. I began by marking lincRNA as lncRNA as this is consistent
with the annotations in HGNC (for the lincRNA's that are in HGNC). In
addition, lncRNA is an ISNDC rna type. However I saw the need to improve
the description for these by using the gene level information. By doing
this I could preserve some of the lincRNA information in the
description. This will now use the gene descriptions for the transcripts
if possible. It will also strip out some weird information that often
gets added to the end of descriptions to make the text prettier.

It would probably be useful to put lincRNA in the product field like
with scaRNA.

In doing updating the code I realized that the interface to lots of
methods needed to change. This commit is the result of that refactoring.
I think the overall design of the code still needs lots of work but this
is a start to that process. For example it would probably be better to
gradually build up an RNAcentral entry as well as have a consistent
interface to all the methods. Right now they are a bit of a mess as can
be seen by the hoops I have to jump through to run the tests.