Improve rna_type detection
Some things have an rna_type of 'misc_RNA'. This is not good because it
is the most uninformative type of annotation. Many of them can be given
an more informative type because they are the results of an Rfam search.
So by inspecting the xrefs for Rfam we can determine the correct
rna_type. I've added changes to do a better job in these cases. It is
not yet comprehensive but I'll inspect all cases later and hopefully
improve the decisions.