Improve the Rfam type detection and begin Mouse
This adds more inferences based upon the Rfam types. I've also started
detailing some of the mouse types. There are many mouse types that are
just predicted gene that I don't know what to do with yet. In addition,
one type 'Nbr1-203' is for a protein. I have to change the code to
better detect if something is ncrna now.