Add importer for entire species/all species
This adds a class that can import a specified species from Ensembl (by
scientific name) or all species at once. The importer will download all
required chromosomal files (unless there are none in which case it will
use non-chromosomal files) via FTP and then process them with either the
generic importer of the GENCODE importer if required. The input species
name must be a string like 'Homo sapiens', not 'human'. It is currently
not documented (and there are some caveats in using it) but that will
be fixed shortly.