Add some utilities for getting Rfam data
This adds some functions to fetch the family information from Rfam. It
will be useful when dealing with Ensembl features which have no
rna_type other than 'misc_RNA'. Some of these features have an
annotation indicating they came from some Rfam model. Thus we can use
the Rfam model's rna_type to infer the correct rna_type. The functions
here help with that.

There is a bit of work needed to infer the correct type as it is not
always in the type field in the dump.