Add a class to download files
This is a better approach than the previous one of not downloading and
keeping the file in memory. First it uses less memory which is nice,
plus when I have to rerun things I already have files downloaded which
makes things much easier. Finally it prevents the disk from being used
by cached instances of these tasks.