Initial work to add a deduplicating stage
It turns out that Ensembl can provide duplicate data split across
several files. In order to deal with that I've added a stage which will
run sort on the command line to do this. It is very simple but sort of
works.

There are some caveats though. We should probably check if multiple
entries the same accession have differing sequences and then select a
'correct' one. Right now we don' have any priority (other than the file
ordering) between entries. This is a potential problem but is a small
one.

Also this uses plumbum but I will try to rewrite without for the next
version.