Improve name to INSDC type logic
Looking at some of the names I found that there are some incorrect hits.
For example, isrP would match the SRP_RNA pattern as it contains 'srp'.
That is incorrect and has been fixed. I've reorganized the code that
fetches the Rfam data a bit so it is cleaner and has better names. I
have also added more tests to cover some of the cases I found by looking
at the matches. Finally, I made the name to INSDC type mapping fail if
the names are not disjoint patterns. This is shouldn't happen but it is
useful to check.