Include old RNA type mapping
I had most of the old RNA type mappings included but I went and copied
all of them in. This makes the guessing procedure more complete. I
should probably consider preventing this from creating so many 'other'
mappings. This is too generic a group to be really useful. Though it may
be best to just wait until RNAcentral uses a more complete system (like
SO terms).