Initial design of code to import Rfam hits
The parser code is a minor modification of the work @nataquinones did
in:

https://github.com/nataquinones/Rfam-RNAcentral/blob/master/toolkit/parser_cmscan.py

The idea behind this is to have a luigi stage that will parse and import
all hits from scanning RNAcentral sequences with Rfam families. This
does not handle importing the metadata about families. The design I'm
leaning towards is to partition the main file by Rfam family so all hits
to a family are in a file. This then stores the data for that family.
it should be possible to parallelize this pretty easily.