Implement initial version of Rfam parsing
This can now parse the .tbl file that is a result of an Rfam scan and
then write a csv file that should be readable by pgloader.

In doing this I moved a lot of code around. Notably, I put the rfam
loading code in rfamf/search_import.py from rfam_hits.py and moved the
utility parsing code into rfam/utils.py. I think this is a clear
organization. In addition, the current loading is assuming that there
will be some SQL to handle filling out the completeness column. This is
easier and much more efficient to do in the database as opposed to
loading everything in python. The resulting is yet to be tested with
pgloader to make sure everything is good however.

I've also fixed a couple bugs along the way.