Improve family loading
We now load much more information about Rfam families than before.
Before we only needed a little information for inference of INSDC types.
Now however we want to use this to also load data into the RNAcentral
database this requires more data.

At the moment it cannot determine what domain a family belongs to (if
any), thus the test for it is commented out.