Initial implementation of family and clan loading
This is my first untested attempt at building classes to load rfam
family and clan data. My goal is to have a single entry point for all
rfam loading, thus the creation of these simple classes. It probably
isn't required to run a whole luigi pipeline but placing it as one will
make it simplier to have a single program to call.