Reorganize Rfam code and add pgloader
The Rfam code is now reorganized to have a straightforward setup and
naming. All tasks that create csv files are named with csv in the name.
Previously these were called 'import' despite not actually importing
anything.

As part of this I extracted some common methods and placed them in a
common class. Notably, all the outputting to csv is now in a single
class that is reused.

In addition, I've added some config classes. I think this is simpler
than using parameters because this provides a single point to configure.
This single point can be documented in one place making it easier to
track down and know of what is required.

In addition, I've started adding tasks that run pgloader control scripts
to the pipeline. The pgloader scripts are all called pgload_*, and there
is a standard base class to build off of. This allows luigi to serve as
the entire way, start to finish, that imports data to the database.

This is all untested so far.