Detect if a family should be suppressed
This will now look at the rna_type to detect if a family should be
suppressed. This information will be used later by RNAcentral when
importing Ensembl and other data sources.

This is done by reimplementing the logic from the perl and sql scripts
that Rfam uses when exporting to RNAcentral. There is likely a better
way to do this.