Use iteration and writerow to write csv's
For some reason if you use writerows with python's csv module it will
try to load all rows into memory (I think). This means that when
processing the hits, instead of using just enough memory to represent
one row we instead try to use enough memory to represent all rows. This
is pretty silly as the amount of memory needed to do that is excessive.