Simplify, standardize and correct output bug
This makes all the PGLoad* classes look more or less the same. In
addition, I've removed the incorrect output method. With this it isn't
simple to just run the PGLoad* tasks because the output is created by
the CSV stages so PGLoad wouldn't run. I've also simplified the code
that gets the filename for pgloader.