Refactor, reorganize and bug fix
This commit completely reworks how the rfam import is structured. The
large set of renames are due to me creating a folder, tasks, for all
luigi tasks to live in. In addition, I've created some utilities that
are specific to luigi tasks and so those become a subdirectory of the
tasks folder.

This tasks folder now has a single entry point that can always be used
to run tasks. I prefer this because we can now clearly define a single
entry point and document how to use it clearly. Doing this lead me to
remove the boiler plate if main from all tasks. This is an organization
I would like to keep.

As part of the refactoring I've moved some common logic into common
places and added more global configuration.

In addition, there a variety of bug fixes for the import procedure. For
example, ensuring that the encoding is correctly set for pgloader and
skipping the correct number of header lines for import rfam hits.