Fix template generation
This adds 2 fixes. First it corrects the template generation so the
pgloader control scripts are actually generated. They were not before
because the db_url parameter was not being filled when it should be.
Secondly, this corrects the work_mem parameter for pgload_hits. It has
to have units 'MB', not 'M'.