Improve rfam data loading
I've recently changed the names of the rfam tables to all start with
rfam_. This required a bunch of renames, which this does. In addition,
instead of manually creating the load_rfam_* tables I will create them
as part of the pgloader process. This way they will only exist while
loading data. In addition, I will know that I didn't forget some key
part of the table, like indexes.