First attempt at storing go terms
This creates a series of tasks to load and store GO terms. Right now all
GO terms are loaded from the Rfam families. It is possible to load all
the data in a single pgloader script but I don't do that for future
expansion. We are likely to need more GO terms than just from Rfam. This
design allows me to add another source fairly easily.