Fisrt pass at parsing GtRNAdb JSON
This is the first attempt at parsing the JSON. It hasn't been checked
to see if it works. I have instead focused on trying to get a good, if
rough mapping between the JSON data and what we will store. I'm sure
there are issues to solve. Notably there is a lot of duplicated code
between this and the Ensembl parses. I will probably merge this, along
with rebasing on top of my ensembl work.