Reorganize and add simple tests
This reorganizes the code a bit. I decided to split up the gtrnadb code
into more than one file just because it was easier for me to follow that
way. There was too much code in one file for my preferences. As part of
this I moved the entry class into a more generic place. I intend to use
this in all the database I parse, including Ensembl.

Secondly, I added some simple tests for the helpers and outlined some
parers tests.