Create initial GtRNAdb importer task
This can create CSV files for importing with the pipeline. I may not
have the result files named correctly yet but the overall logic is
correct and functioning.

However, in doing this I discovered a problem. GtRNAdb uses some
outdated taxon ids. This is a problem since we try to normalize
phylogeny and the like. Because they are oudated our current method
fails. It is possible to get the updated taxon id with uniprot (and
possibly NCBI as well). It isn't clear who, if anyone, should handle
this issue.

In addition, this contains a lot of code duplication. The ensembl
writers are basically a copy of the writers I'm adding here. These are a
bit nicer to work with and are what I will use in the future. Once this
branch and the Ensembl branch have been merged a clean up must be done.