Allow for 500's
Sometimes the taxonomy services gives 500's. Retrying the request will
fix this. So to deal with this I dd some logic to retry requests if
needed. This seems to work ok, but sometimes it will still have 500's. I
will look to another service to replace the ENA one but finding a good
one is tricky so far.