Initial implementation of secondary structure load
This is my first attempt at loading secondary structures. This adds a
utility class for running the pgloader scripts and then tasks for
running the import on all files in the secondary structure directory, or
just selected files. This is because we may want to import selected
databases one at a time. This is untested currently but should cover the
majority of what needs to happen.

The known issue is that this will always update accessions even when
inserting a duplicate secondary structure. In reality we should only
update if the secondary structure has changed. This will lead to extra
inserts which could slow the process down. Sadly, it isn't possible to
add two ON CONFLICT statements so a simple method of deal with this is
out. It is possible to add some if statements but I doubt that will
really speed things up. If speed becomes an issue this can be explored.

I also have not made GtRNAdb import depend on this import. Maybe I will
later but currently I think it is not the correct approach as we want to
import secondary structures after the standard pipeline and the standard
pipeline is currently controlled outside of luigi.