Bug fixes
This fixes the bugs that prevented the import from running. This will
now correctly import secondary structure data into the database. The
current downside to this approach is that since the secondary structure
files may contain duplicates we have to use a select distinct which
could lead the insert being very expensive in some future where we have
many secondary structures to insert.