Correct GENCODE parser
When parsing a GENCODE file we should treat every accession in the file
as a GENCODE entry. This is a change from before where we could detect
the OTT* entries. This is no longer true as they have been removed. Thus
we now treat all entries as being from GENCODE. I checked chromosome 1
and this is a valid assumption.