Add tasks to fetch MGI data
The MGI data requires a bit of parsing after downloading. For that
reason I've added some tasks here to handle downloading the data and
dealing with some issues. We are going to do the import via Django but
I'm going to keep a lot of the logic for the more generic parts, like
inferring RNA type here.