Add a cleanup task
We need to cleanup data as well as load it. To do this I've added a
cleanup task that does not depend on any other tasks. This is because it
must be run manually after we check that the release has been loaded
properly.

I've also reworked the cursor and such so that we no longer need
annoying multiple inheritance. Plus I've add the needed imports to the
top level task module.