Reorganize and reduce code
There was at least one level of inheritance too much for the
load_accessions, references, etc tasks. I've removed that so the code is
simpler. In addition, loading secondary structures is part of a
release. So I've moved the code into the release directory and
simplified it as well. In addition, I've merged LoadShort and LoadLong
into LoadSequences. We generally have no desire to load only long or
short sequences, but it is still possible, so I created a single task
for it.