Initial attempt at adding search export
Currently there is some search export logic in the web application. This
may not be the best place for it. So I'm trying out adding it to the
pipeline code as it may be better here. It would be easier to create the
correct number of jobs and such.

This is also a very different approach than in the search export. This
approach is based off building a data structure to represent the XML and
then building the ElementTree objects which are then turned into XML and
written out. The previous method builds a lot of strings basically.

The code isn't yet tested or run but is an outline of what I think I
will do. I'm saving this now because it is a fair amount of work and can
be worked on further.