Alter Ensembl external id
This alters the external id that we generate for ensembl to lack the
version string. This is because the versionless id is generally present
in other databases so querying for this id is faster this way. This
doesn't change the accession as we need the accessions to change if the
sequence changes.

In addition, I've also made a few small changes to how the ensembl
entries are computed. I'm likely to make more in the future to move to
the simple style that I use for other databases.