First attempt at using Entry and EntryWriters
I want to move to using my entry writers and the like because they
produce a single Output object which behaves much better in luigi.
Notably it should not produce many duplicate outputs unlike the current
setup does sometimes. This is my first attempt at moving the Rfam work
to use this setup. It is currently untested but this is the start.