Make seq_version required
A seq version must be set to a non-empty string in order to create a
valid Entry. This is because it is required in the database for
correctly working with xrefs. I made a mistake and allowed it to to be
null in the db and that caused all manner of problems. Hopefully this
works toward preventing that next time.