Ensure that there is always a chunk_00 file
In some cases, like the long sequences for GtRNAdb, there will be no
files produced by split because it has no long sequences. This leads to
an empty merged.csv which leads to no chunk_00 file. This causes the
loading task to fail. We don't want this so instead we check if
merged.csv is empty in which case we produce an empty chunk_00 file.