Update GtRNAdb to use latest format
They now provide a mature_sequence field which we prefer over sequence.
In addition, they provide a better description to sequences. Thus we use
these newer entries when possible.

In addition, we will not assign a 2D to a sequence if their lengths
differ.

This implements that as well as providing a series of tests.