Add a task to export Rfam annotations
This will export all Rfam annotations on active xrefs. An important part
of this is finding which xrefs are active. To do so I create a temp
table called rna_active to store the active UPI's. To build this I
insert from each *_not_deleted table. This seems to be the quickest way
to get this data. This is a bit hacky, but does seem to work quickly.