First attempt at FASTA export
This isn't perfect yet. There are issues where the connection to the db
can time out, which is bad. However this is close enough to correct to
commit a version of. The nhmmer related code does work, which is good.
More work is needed to get the active and inactive export to always
work.