Correctly using copy command
This will correctly use the copy command to write to a temp file. This
file is then used processed to generate a real FASTA file once copy is
done. This moves code around so we only have the connection and final
file open when we need them. In addition, we correctly open and close
the temp file. The file has to be written, close and then opened again
for reading in order for this to work.