Initial ENA parser
This is a very simple ENA parser. It can correctly parse a single
example that I picked at random. This does not include all the data that
could be included. This will be extended as I find more examples to work
with.

In addition, I've started moving around the generic helpers for working
with EMBL files. We also parse EMBL files from Ensembl so I have some
common functions that I could move around. I probably broke Ensembl
parsing for the moment, but I will add that back later.