Do a better job with anticodons and notes
This can now parse anticodon data and note data in a cleaner way. This
will do a better job of actually determining the anticodon sequence in
some cases. It also produces a more useful note in that it will pull out
the ontology terms so it is possible to do something useful with them
later.

Pleasingly the biopython parser is better than the bioperl parser
in that it can handle mistakes in the format. This is tested by parsing
all the problem files anton has added.