Add tasks for importing ENA data
The goal is to not produce tens of thousands of tasks for reading in ENA
data files. The issue with that is that many of the files are small and
thus the overhead of each task will be very large. Instead we try to do
things in a reasonable amount of tasks.

This is done by having a single task for each wgs/* directory and then a
single task for each con/*, std/*, etc. The con, std and so forth have
not too many (maybe a hundred files) so a single task per each isn't too
bad. On the other hand the wgs directory has a several sub-directories
with many files in each. This provides a natural grouping to each one
and makes for an easy way of grouping the work.

If this still produces too much overhead then I will probably just cat
lots of files together.