Be smarter about bad taxids
Sometimes we cannot look up the correct tax ids from ENA. Not sure why
that happens but until we have a better solution (normalized tables and
load taxonomies from Uniprot) we will use this approach. Basically we
fall back to what is in the file if we cannot get something newer. This
should allow us to handle nearly all ENA files.