Store ensembl locus tag in gene and locus tag
This is something Anton and I discussed a while ago. We do not lose data
doing this as we still have gene id in the optional_id. I know this was
an improvement we wanted, but the gain is unclear right now.

We should really do a better job keep data stored in a consistent and
clear way.