Begin work on importing using the JSON schema
This is the start of work to use the JSON schema to import data. This
does not yet run but is an outline of what this approach will look like.

Writing this is making me think we need to restructure our data quite a
bit. Both in the DB and the pipeline. For example, we should have a
separate step to import and populate the references (look up data using
Europe PMC) as well as normalize the taxonomic data.