Rework to use psql and cleaner config
The code is a bit cleaner now which is nice. I have also refactored the
pipeline to use psql in place of the python connection. This still fails
in Docker and I am unclear why as this seems to work fine on the
cluster. I am going to try running this in the cluster and hope it works
there. That is very much not ideal but better than nothing.