Improve handling of some ENA data
This does a better job of extracting function, pseudogene, anticodon and
operon data. It can now pull more of that information out of files that
previously. In addition it can find new publications that are stored in
the experiment field and use Europe PMC's API to fetch the publication
data.