Begin work on parsing out TPA's
We should parse the TPA mapping file to only produce output for TPA
entries or the ENA entries without TPA's, not both. This tarts that work
by adding some basic logic for parsing a TPA file.

I've also started working on creating a mapping and producing mapped
entries as needed. The approach will be to store all entries in a simple
dict in memory and use that to find the TPA data. If memory becomes a
problem I may stick things in some simple DB like sqlite, but I would
prefer to keep things in memory.