Apply TPA mappings
This can apply TPA mappings in a fairly reliable way. It has some logic
to deal with cases like WormBase where the parent_accession is a
chromosome and so the mapping must rely upon the locus_tag as well as
the parent_accession. This should provide a testable and possibly
correct implementation of loading TPA mappings.