Add more ENA import tasks
This will now copy files to a more local location before trying to
import. This has proven key to debugging in the past, thus we implement
it here. There are also several changes to support this in various
places like the config.

In addition, we download the TPA files and map and import data. The
mapping will mean we do not produce ENA entries for things that come TPA
entries, which is what we want.