Begin work on moving pdb import to pipeline
All import logic should be in the pipeline to ensure we have a simple,
clean way to running the pipeline. This begins the work on moving the
logic into the pipeline by outlining where all the code will go. The
code is mostly cut and paste from the current script, but is untested. I
need to download some test data and implement the tests for this to be
ready.