Rework utils and test
This adds tests to the fasta export utils as well as making the code
more testable. The changes are just to support easier testing. I also
allowed the code to find the path to psql using distutils. Seems in my
setup that doesn't always work correctly, so this deals with that.