Add tests for counts exported
This adds some simple tests to ensure the number of entries found for
fasta active and inactive export match the expected counts. Currently,
the tests indicate that we should have more sequences (13,729,588) than
are in our current text index (13,353,531). I trust the test counts
since there is a clear table with the active upi's in it. I think the
export procedure should use that as that would produce consistent counts
at least.