Move id_mapping code around
The id mapping code should be organized like all other code where the
actual logic is kept separate from the code that setups tasks and the
like. This change does that. In addition, we have some tests for the
id_mapping work so we can be sure it probably works.