Attempt to fix WormBase and PomBase issues
Those databases do not produce the correct counts of results after
mapping and importing the data. This attempts to fix those issues by
improving the logic used to produce the mapping keys as well as what
data is stored in a TPA. It seems despite there being a 'standard' way
to represent TPA's and associated data ENA doesn't actually follow that
standard.

As part of it I've added some data files and more tests.