Add functions to fetch GtRNAdb links
GtRNAdb provides a page that lists out the locations of the data files
to download. This adds functions to fetch that summary page and then
parse out the locations of the urls of data files to download. This
involves some less than ideal work of parsing HTML via regexs. As this
is the only place we handle HTML I am ok with leaving it like this.

This will be used to create a task that downloads data files so GtRNAdb
work can involve no manual work.