Add utilities for dealing with compressed files
This adds a utility to expand compressed files. It will handle tarballs
and gzipped files. It provides a method of autodetecting the type of
file and expanding it as needed. In addition, the expansion will create
a list of the files that were expanded. This is useful when with
tarballs as there may be many files contained and this provides a useful
way of expanding and then dealing with each file.