Make GtRNAdb import completely self contained
This no longer requires human intervention to download data. GtRNAdb
will now download and unpack data as needed. This is one of our goals
with the pipeline, to have the entire process be self-contained and need
no human intervention.

Doing this adds several utilities to make fetching and downloading data
easier.