More work on importing via json schema
This adds more code to import via a JSON schema. As the JSON schema is now
in a probably usable state, at least for FlyBase, I am now adding some
tests. The test focus on FlyBase data. Not all tests pass now this now
has code that actually runs. It seems that we are getting more (gene
information specifically) and less (missing product and other minor
things) so the results may be different. Notably we now have transcript
ids so we can and should use those in places like accessions and
external id. For this reason the existing data is not exactly reproduced
but it is very similar.