Add tests for RGD import
This tests that the RGD code can at least process some of the basic
data. This has some limitations. Notably, it assumes all data is Rat
(but they have more than just rat) and it does not yet extract
sequences. However, this does appear to cover the basics of process RGD
data.