Begin reworking to use provided fasta file
RGD will now be providing a fasta file that contains the sequences to
use. This means we no longer need to use gff files and chromosome fasta
files to extract sequences. This allows me to simplify several parts of
the pipeline. I keep the same general idea of producing a fasta file and
then indexing it to find sequences.

A little work needs to be done to the file that RGD provides as they do
not use the gene id as the id, which is what I want to use. They seem to
use an genbank id. If using the gene id fails I suppose we can use that
instead.

I have outlined some tests but this is not complete yet. All tests will
fail because I haven't added test data to the fixtures or updated the
function calls.