Test RGD parsing
This provides a more complete set of tests to see if this can correctly
parse RGD files and produce reasonable data. The data is limited at the
moment (missing publications, sequences only having 1 exons) but seems
to be enough to start with. Fetching publications will have to wait
until I have merged the ENA code into master and can rebase as that
branch has the logic needed to fetch publication information from pmid.