Try to correctly deal multi locus genes
This will map to all sequences in the given fasta file. This produces
entries which refer to the same gene but may have different sequences.
Most importantly this will now correctly parse the locations out of the
tsv file if there are several locations. This will produce a single gene
which may have more than one sequence in several places. The sequences
in a single locus may differ as well. There is no check made to ensure
that the sequence of the given location matches the reported sequence.