Work to parse miRBase optional ids
This can now extract the optional id by parsing DR lines from EMBL
files. This requires parsing the EMBL file once to get a mapping of DR
information, then a second time to actually build the data. The
optional_id is only applied when we determine what TPA applies to a
mapping. Until then the information is stored in the xref_data field as
it is an xref.

Additionally, I corrected some failing tests and maybe helped with the
unicode issue. Really I should sit down and figure that out properly at
some point.