Refactor and being SILVA import
This doesn't actually do anything that is required to import SILVA data
but does start the work to do so. Additionally, I have refactored the
code to hopefully be a little simpler and easier to extend. This
involves some renaming, as well as storing the extended data that ENA
can provide.

Also, I've renamed `mapping` to `tpa` to make it clearer what the code
is working with.