Add a ProcessData task
While being able to run import using a single database is great for
testing it is a pain for releases. This adds a single entry point to
handle processing (turning raw data into CSV's) for all databases that
we import on each release. Ideally every database would be here, but
some are only computed using Django models and others may not be
reliable enough to go into this list.