Create a new approach to search export
This approach to search export is based upon doing more work in the
database to turn the rows into a grouped data structure. While this is
expensive it seems to go fairly fast. The overall design is we use the
database to produce a single JSON object that represents everything we
need to know about a entry. Happily, pyscopg2 supports json data
structures natively so we get a python dict out of the query, which is
great. We then process this dict to produce the XML data structure which
is then written out.

I'm drinking the functional programming juice pretty hard right now,
thus the reason this is designed the way it is. However, it is pretty
easy to work with a series of small functions that produce The data we
want.

The tests do not all pass, yet. These were some sequences I selected a
while ago and are probably worth testing out. I need more comprehensive
tests that check the entire data structure is correct for a few
important cases, rather than the current set which just checks if a
particular value is correct.