Attempt to use rnc_genome_mapping for coordinates
We are starting to infer genomic locations and populating a table of
such data. We want the search export to include such information so we
modify the query and the xml generation to include this data. Basically
we will check if we have either a given or inferred coordinate and if so
produce 'True' as the xml value.