Attempt to create the release note
This should generate the release note by reparsing the produced XML and
creating the note. It doesn't set the release number but maybe it
should. Additionally, it reparses the xml because that this the safest
way to pull out the data. The writer now puts everything on one line, so
grep is a bit trickier. Sadly this is probably very expensive.