Make query for efficient
This makes the query easier on the database. First, we use the
has_coordinates column on rnc_rna_precomputed instead of computed this
flag dynamically. This makes the query faster and prevents excessive
memory usage. Previously we were getting some out of memory errors probably
becaause of the rnc_genome_mapping table.

Additionally, this changes the (array_agg(column))[1] to just
array_agg(column). This will no longer pull the unique value from the
database which may also ease some strain on the db. This is pretty minor
so maybe it wasn't that useful.